refactor gamebrowser from landing module to its own module
