+++ frontend/src/Router.js
@@ -8,7 +8,7 @@ import App from "./modules/app/";
 import Profile from "./modules/profile/";
 import Settings from "./modules/settings/";
 import { loadUser } from "./modules/app/actions";
-import GameBrowser from "./modules/landing/components/gamebrowser"
+import GameBrowser from "./modules/game-browser"
 
 const notFound = () => {
   return <p>Not Found</p>;
