+++ frontend/src/modules/game-browser/actionTypes.js
@@ -1,0 +1,2 @@
+export const GET_GAME_DATA = 'GET_GAME_DATA';
+export const GET_POPULAR = "GET_POPULAR";
