+++ frontend/src/modules/game-browser/components/filters/genres.js
