+++ frontend/src/modules/game-browser/components/filters/index.js
