+++ frontend/src/modules/game-browser/index.js
@@ -1,18 +1,11 @@
 import React from "react";
 import "./styles.css";
-import { getPopular, getGameData } from "../../actions";
+import { getPopular, getGameData } from "./actions";
 import { connect } from "react-redux";
-import Moment from "react-moment";
 import { Button, Card, Label, Dropdown } from "semantic-ui-react";
-import { Backdrop } from "../../../app/components";
-import { TileCoverLoader, TextLoader } from "../../../game/components";
-import { GenreFilter, DateFilter, DeveloperFilter } from "../filters";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  faXbox,
-  faPlaystation,
-  faSteam,
-} from "@fortawesome/free-brands-svg-icons";
+import { Backdrop } from "../app/components";
+import { GenreFilter, DateFilter, DeveloperFilter } from "./components/filters";
+import GameTile from "./components/gametile";
 
 const pickers = [
   { key: "genre", text: "Genre", value: "Genre" },
@@ -55,79 +48,6 @@ class GameBrowser extends React.Component {
       });
     }
   }
-
-  renderTile = (game, key) => {
-    const getDeveloperName = companies => {
-      if (companies) {
-        var dev = companies.find(c => {
-          return c.developer === true;
-        });
-        if (dev) return dev.company.name;
-      }
-      return "Not Found";
-    };
-
-    const getPlatformIcons = platforms => {
-      return (
-        <React.Fragment>
-          {platforms.map(platform => {
-            let icon = null;
-
-            if (platform.name.includes("PC")) icon = faSteam;
-            else if (platform.name.includes("Playstation"))
-              icon = faPlaystation;
-            else if (platform.name.includes("Xbox")) icon = faXbox;
-            return <FontAwesomeIcon icon={icon} size={"sm"} />;
-          })}
-        </React.Fragment>
-      );
-    };
-    if (game) {
-      const description = game.summary
-        ? game.summary.slice(0, 100) + "..."
-        : "";
-
-      const cover = game.cover
-        ? `https://images.igdb.com/igdb/image/upload/t_cover_big/${
-            game.cover.image_id
-          }.jpg`
-        : null; //TODO there should be a default cover if the API doesnt provide one
-
-      const developer = game.involved_companies
-        ? getDeveloperName(game.involved_companies)
-        : null;
-
-      const date = game.first_release_date ? (
-        <Moment format="YYYY">{game.first_release_date * 1000}</Moment>
-      ) : null;
-
-      const header = game.name ? game.name : null;
-
-      const platforms = game.platforms ? game.platforms : [];
-
-      return (
-        <Card
-          className="game-card"
-          image={cover}
-          header={header}
-          meta={developer && date}
-          description={description}
-          href={"/games/" + game.slug}
-          key={key}
-          extra={getPlatformIcons(platforms)}
-        />
-      );
-    } else {
-      return (
-        <Card
-          className="game-card"
-          image={TileCoverLoader}
-          description={TextLoader}
-          key={key}
-        />
-      );
-    }
-  };
 
   removeFilters = (key, filter) => {
     let filters = { ...this.state.filters };
@@ -267,7 +187,7 @@ class GameBrowser extends React.Component {
           {this.renderFilters()}
           <div className="game-grid">
             {this.state.filteredGames.map((game, index) => {
-              return this.renderTile(game, "game" + index);
+              return <GameTile game={game} key={"game" + index} />;
             })}
           </div>
           <div className="center">
@@ -285,7 +205,7 @@ class GameBrowser extends React.Component {
             {[...Array(20).keys()].map((val, index) => {
               return (
                 <div className="game-box" key={"game" + index}>
-                  {this.renderTile(null)}
+                  <GameTile game={null}/>
                 </div>
               );
             })}
@@ -297,9 +217,8 @@ class GameBrowser extends React.Component {
 }
 
 const mapStateToProps = state => ({
-  popular: state.landing.popular,
-  loading: state.landing.isLoadingPopular,
-  games: state.landing.gameData
+  popular: state.gameBrowser.popular,
+  games: state.gameBrowser.gameData
 });
 
 export default connect(
