+++ frontend/src/modules/game-browser/reducers.js
@@ -1,10 +1,8 @@
-import { GET_POPULAR, GET_BACKDROP, GET_GAME_DATA } from "./actionTypes";
+import { GET_POPULAR, GET_GAME_DATA } from "./actionTypes";
 
 const initialState = {
-  backdrop: {},
   popular: [],
-  gameData: [],
-  isLoadingPopular: true
+  gameData: []
 };
 
 export default function(state = initialState, action) {
@@ -15,11 +13,13 @@ export default function(state = initialState, action) {
         popular: action.payload,
         isLoadingPopular: false
       };
-    case GET_BACKDROP:
+    case GET_GAME_DATA: {
+      const gameData = [...new Set([...state.gameData, ...action.payload])]; //concat arrays and remove duplicates
       return {
         ...state,
-        backdrop: action.payload
+        gameData: gameData
       };
+    }
     default:
       return state;
   }
