+++ frontend/src/modules/game-browser/styles.css
