+++ frontend/src/modules/landing/actions.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { GET_POPULAR, GET_BACKDROP, GET_GAME_DATA } from "./actionTypes";
+import { GET_POPULAR, GET_BACKDROP} from "./actionTypes";
 
 export const getPopular = (limit = 6, offset = 0, filters={} ) => dispatch => {
   axios
@@ -13,23 +13,7 @@ export const getPopular = (limit = 6, offset = 0, filters={} ) => dispatch => {
     .catch(err => console.log(err));
 };
 
-export const getGameData = (gameSlugs) => dispatch => {
-  //break the array into chunks of 10 and perform a request for each chunk
-  while(gameSlugs.length > 0){
-    const slugs = gameSlugs.slice(0,10)
-    gameSlugs = gameSlugs.slice(10)
-    axios.get("/api/games/multi", {
-      params: {
-        slugs: slugs.join(",")
-      }
-    }).then(res => {
-      dispatch({
-        type: GET_GAME_DATA,
-        payload: res.data
-      })
-    })
-  }
-}
+
 export const getBackdrop = gameId => dispatch => {
   axios.get(`/api/games/backdrop/${gameId}/`).then(res => {
     const data = res.data[0];
